backend/requirements/base.txt
# backend/requirements/base.txt
# Core dependencies for production

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
starlette==0.27.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph with API support
langgraph==0.2.16
langgraph-sdk==0.1.13
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.28
langchain-openai==0.1.20

# Local LLM integration
ollama==0.3.0
transformers==4.44.0
torch==2.1.0
tokenizers==0.19.1

# Vector database and embeddings
chromadb==0.5.5
sentence-transformers==3.0.1

# HTTP client for API calls
httpx==0.27.0
aiohttp==3.10.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.2.0
cryptography==42.0.8

# JWT handling
pyjwt==2.8.0
authlib==1.3.1

# Session management
itsdangerous==2.2.0

# WebSocket authentication
websocket-client==1.7.0

# Environment and configuration
python-dotenv==1.0.0

# Database for user management and session storage
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # For PostgreSQL async
aiosqlite==0.20.0  # For SQLite async

# Redis for session storage and caching
redis==5.0.1
aioredis==2.0.1

# Async support
aiofiles==23.2.1

# Code execution sandbox
subprocess32==3.5.4; sys_platform != "win32"

# Logging and monitoring
loguru==0.7.2
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9
limits==3.10.1

# CORS handling
fastapi-cors==0.0.6

# API versioning
fastapi-versioning==0.10.0

python-jose[cryptography] 
passlib[bcrypt] 
authlib